work on introduction.
now it is simply too long. we might need some subsections, like the
prior "essence of julia" section though perhaps renamed.
the segue to the paragraph that mentions pypy is awkward.
our gripes are now in two parts: the user-facing problems with the
two-tier thing, and the need to move domain-specific functionality
to libraries. i think we are aware of too many problems, and
it becomes a litany. we should limit to mentioning the most hard-hitting
problems: (1) usability. it's hard to program in 2 languages. (2) technical.
the resulting systems are hard to compile well, and more expressivity and
genericity is needed to write good libraries. something like that.